refactor: simplify flight server exceptions
Make sure the exception types are ordered properly.
Also fix the CLI help text.

JIRA: CQ-764
risk: low
