+++ gooddata-flight-server/tests/utils/test_methods_discovery.py
@@ -1,12 +1,7 @@
 #  (C) 2024 GoodData Corporation
 
 import pytest
-from gooddata_flight_server.exceptions import (
-    FlightMethodsModuleNotFoundError,
-    InvalidFlightMethodsFactoryError,
-    MultipleFlightMethodsFactoriesError,
-    NoFlightMethodsFactoryError,
-)
+from gooddata_flight_server.exceptions import FlightMethodsModuleError
 from gooddata_flight_server.server.base import FlightServerMethods
 from gooddata_flight_server.utils.methods_discovery import get_methods_factory
 
@@ -26,7 +21,7 @@ def test_get_methods_from_invalid_module_multiple():
     """
     Test the methods discovery function on a mock module that provides more than one methods factory.
     """
-    with pytest.raises(MultipleFlightMethodsFactoriesError):
+    with pytest.raises(FlightMethodsModuleError):
         get_methods_factory(".multiple_methods_module", root="tests.utils")
 
 
@@ -34,7 +29,7 @@ def test_get_methods_from_invalid_module_none():
     """
     Test the methods discovery function on a mock module that provides no methods factories.
     """
-    with pytest.raises(NoFlightMethodsFactoryError):
+    with pytest.raises(FlightMethodsModuleError):
         get_methods_factory(".no_methods_module", root="tests.utils")
 
 
@@ -42,7 +37,7 @@ def test_get_methods_from_invalid_module_invalid_signature():
     """
     Test the methods discovery function on a mock module that provides an invalid methods factory.
     """
-    with pytest.raises(InvalidFlightMethodsFactoryError):
+    with pytest.raises(FlightMethodsModuleError):
         get_methods_factory(".invalid_methods_module", root="tests.utils")
 
 
@@ -50,5 +45,5 @@ def test_get_methods_from_invalid_module_non_existent():
     """
     Test the methods discovery function on module that does not exist.
     """
-    with pytest.raises(FlightMethodsModuleNotFoundError):
+    with pytest.raises(FlightMethodsModuleError):
         get_methods_factory(".non_existent_module", root="tests.utils")
